+++ src/EngineGUI/HierarchyPanel.cpp
@@ -34,7 +34,7 @@ void NoxEngineGUI::updateHierarchyPanel(NoxEngine::GameState* state, GUIParams *
 		// TODO: revisit example if we want multi-selection
 
 		// Available panel width
-		int width = ImGui::GetContentRegionAvail().x;
+		f32 width = ImGui::GetContentRegionAvail().x;
 
 		for (int i = 0; i < state->activeScene->entities.size(); i++) {
